Feature: error snackbar (#18)
* Split out Results from Search

* Snackbar for errors, some refactoring

* Refactor all material imports, reduces module size
