+++ src/theme.tsx
@@ -1,7 +1,5 @@
 
-import React from 'react';
-
-import { createTheme } from '@mui/material';
+import { createTheme } from '@mui/material/styles';
 import darkScrollbar from '@mui/material/darkScrollbar';
 
 export const tgTheme = createTheme({
