+++ src/workbench/Banner.tsx
@@ -1,7 +1,10 @@
 
 import React from 'react';
 
-import { Typography, Box, Stack, CircularProgress } from '@mui/material';
+import Typography from '@mui/material/Typography';
+import Box from '@mui/material/Box';
+import Stack from '@mui/material/Stack';
+import CircularProgress from '@mui/material/CircularProgress';
 
 import { useProgressStateStore } from './state/ProgressState';
