+++ src/workbench/CenterSpinner.tsx
@@ -1,7 +1,8 @@
 
 import React from 'react';
 
-import { Box, CircularProgress } from '@mui/material';
+import  Box from '@mui/material/Box';
+import CircularProgress from '@mui/material/CircularProgress';
 
 import { useProgressStateStore } from './state/ProgressState';
