+++ src/workbench/chat/ChatConversation.tsx
@@ -1,7 +1,8 @@
 
 import React from 'react';
 
-import { Card, CardContent } from '@mui/material';
+import Card from '@mui/material/Card';
+import CardContent from '@mui/material/CardContent';
 
 import ChatHistory from './ChatHistory';
 import InputArea from './InputArea';
