+++ src/workbench/entity/EntityNode.tsx
@@ -1,7 +1,7 @@
 
 import React from 'react';
 
-import { Button } from '@mui/material';
+import Button from '@mui/material/Button';
 
 import { useWorkbenchStateStore } from '../state/WorkbenchState';
 import { Value } from '../state/Triple';
