+++ src/workbench/entity/LiteralNode.tsx
@@ -1,7 +1,7 @@
 
 import React from 'react';
 
-import { Typography } from '@mui/material';
+import Typography from '@mui/material/Typography';
 
 import { Value } from '../state/Triple';
