+++ src/workbench/load/Url.tsx
@@ -1,7 +1,8 @@
 
 import React from 'react';
 
-import { Box, TextField } from '@mui/material';
+import Box from '@mui/material/Box';
+import TextField from '@mui/material/TextField';
 
 interface UrlProps {
     value : string,
